nEARLY FINISHED. oDD SERC IMPORTING BEHAVIOUR
