+++ app/Console/Commands/CompetitionImport.php
@@ -3,6 +3,7 @@
 namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
+use function Laravel\Prompts\info;
 use function Laravel\Prompts\text;
 
 class CompetitionImport extends Command
@@ -31,8 +32,20 @@ class CompetitionImport extends Command
     {
         $filePath = "/home/noah/L2023.xlsm";
 
-        $importer = new \App\Importer\Importer($this);
-        $importer->import($filePath);
+        $loader = new \App\Importer\Loader($this);
+        $data = $loader->load($filePath);
+
+        info("All events loaded. Starting import");
+
+        $competitionDetails = [
+            'name' => text('What is the name of the competition?'),
+            'when' => text('When was the competition?')
+        ];
+
+        $importer = new \App\Importer\Importer();
+        $comp = $importer->import($competitionDetails, $data['teams'], $data['speeds'], $data['sercs']);
+
+        info("View at: " . route('comps.view', [$comp->id]));
 
         return Command::SUCCESS;
     }
